+++ lib/commercetools-api/src/Models/ApprovalRule/ApprovalRuleSetCustomTypeAction.php
@@ -1,0 +1,47 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file has been auto generated
+ * Do not change it.
+ */
+
+namespace Commercetools\Api\Models\ApprovalRule;
+
+use Commercetools\Api\Models\Type\FieldContainer;
+use Commercetools\Api\Models\Type\TypeResourceIdentifier;
+use Commercetools\Base\DateTimeImmutableCollection;
+use Commercetools\Base\JsonObject;
+
+interface ApprovalRuleSetCustomTypeAction extends ApprovalRuleUpdateAction
+{
+    public const FIELD_TYPE = 'type';
+    public const FIELD_FIELDS = 'fields';
+
+    /**
+     * <p>Defines the <a href="ctp:api:type:Type">Type</a> that extends the ApprovalRule with <a href="ctp:api:type:CustomFields">Custom Fields</a>.
+     * If absent, any existing Type and Custom Fields are removed from the ApprovalRule.</p>
+     *
+
+     * @return null|TypeResourceIdentifier
+     */
+    public function getType();
+
+    /**
+     * <p>Sets the <a href="ctp:api:type:CustomFields">Custom Fields</a> fields for the ApprovalRule.</p>
+     *
+
+     * @return null|FieldContainer
+     */
+    public function getFields();
+
+    /**
+     * @param ?TypeResourceIdentifier $type
+     */
+    public function setType(?TypeResourceIdentifier $type): void;
+
+    /**
+     * @param ?FieldContainer $fields
+     */
+    public function setFields(?FieldContainer $fields): void;
+}